Em MemoriaControle: --> Iniciando a popular os vetores. Para isso foi criada uma função chamada Cria_Vetor. Você a chama, e os parametros são os comandos da ULA, MUXs relacionados, e os operadores fonte e destino. Se alguma coisa não for usada na micro, dele ser passado NULO ou NULOV(Nulo para Vetor)
Nas outras classes:
--> Diversas correções no código, coisas pequenas.

+++ Arquitetura/src/br/ufrj/dcc/arq/modelo/Multiplex2E.java
@@ -16,8 +16,8 @@ public class Multiplex2E {
 	public static final int L = 0;
 	
 	/**  Configurações dos sinais do multiplex e as operações correspondentes */
-	public static final int[] PASSAMEM  = {0};
-	public static final int[] PASSASULA = {1};
+	public static final short[] PASSAMEM  = {0};
+	public static final short[] PASSASULA = {1};
 	
 	/** Recebe valores Lidos da memória. */
 	public short a;
